feat : detach observer(userClicked) from controller
