+++ HolubLife/src/com/holub/life/model/Neighborhood.java
@@ -325,7 +325,7 @@ public final class Neighborhood implements Cell, Observable {
      * Modifies activeEdges to indicate whether the addition
      * of the cell at (row,column) makes an edge active.
      */
-    private void rememberThatCellAtEdgeChangedState(int row, int column) {
+    public void rememberThatCellAtEdgeChangedState(int row, int column) {
         if (row == 0) {
             activeEdges.add(Direction.NORTH);
 
@@ -367,21 +367,7 @@ public final class Neighborhood implements Cell, Observable {
      * Notification of a mouse click. The point is relative to the
      * upper-left corner of the surface.
      */
-    public void userClicked(Point here, Rectangle surface) {
-        int pixelsPerCell = surface.width / gridSize;
-        int row = here.y / pixelsPerCell;
-        int column = here.x / pixelsPerCell;
-        int rowOffset = here.y % pixelsPerCell;
-        int columnOffset = here.x % pixelsPerCell;
 
-        Point position = new Point(columnOffset, rowOffset);
-        Rectangle subcell = new Rectangle(0, 0, pixelsPerCell,
-                pixelsPerCell);
-
-        grid[row][column].userClicked(position, subcell); //{=Neighborhood.userClicked.call}
-        amActive = true;
-        rememberThatCellAtEdgeChangedState(row, column);
-    }
 
     public boolean isAlive() {
         return true;
